// src/pages/explorer/components/OpcUaServer/data.d.ts
export interface OpcUaServerDataType {
  id: number;
  endpointUrl: string;
  serverName: string;
  fullName: string;
  opcUaProtocolId: number;
  securityPolicyUri: string;
  securityMode: string;
  supportedFile: string;
  authenticationTypes: string;
  productUri: string;
  applicationUri: string;
  applicationName: string;
}